fix(token): 修复启动系统时令牌没有正常加载的问题

+++ src/main/java/plus/lgx/ordermanager/service/UserService.java
@@ -17,5 +17,5 @@ import java.util.Map;
  */
 public interface UserService extends IService<User> {
 
-    Map<Integer, UserModel> getPerms(HashSet<Integer> ids);
+    Map<Long, UserModel> getPerms(HashSet<Long> ids);
 }
